requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
astunparse==1.6.3
cachetools==4.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
gast==0.3.3
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.8
jsonpatch==1.24
jsonpointer==2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.3.0
numpy==1.16.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.2.1
Pillow==5.3.0
protobuf==3.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.12.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.12.0
tensorflow-cpu==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
torchfile==0.1.0
tornado==6.0.3
urllib3==1.25.6
visdom==0.1.8.9
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.12.1
